+++ membership-token/program/src/state.rs
@@ -1,0 +1,95 @@
+//! Module provide program defined state
+
+use crate::utils::STRING_DEFAULT_SIZE;
+use anchor_lang::prelude::*;
+
+#[account]
+pub struct Store {
+    pub admin: Pubkey,
+    pub name: String,
+    pub description: String,
+}
+
+impl Store {
+    pub const LEN: usize = 8 + 32 + STRING_DEFAULT_SIZE * 4 + STRING_DEFAULT_SIZE * 4;
+}
+
+#[derive(AnchorDeserialize, AnchorSerialize, Clone, Debug, PartialEq, Eq)]
+pub enum SellingResourceState {
+    Uninitialized,
+    Created,
+    InUse,
+    Exhausted,
+    Stopped,
+}
+
+#[account]
+pub struct SellingResource {
+    pub store: Pubkey,
+    pub owner: Pubkey,
+    pub resource: Pubkey,
+    pub vault: Pubkey,
+    pub vault_owner: Pubkey,
+    pub supply: u64,
+    pub max_supply: Option<u64>,
+    pub state: SellingResourceState,
+}
+
+impl SellingResource {
+    pub const LEN: usize = 8 + 32 + 32 + 32 + 32 + 32 + 8 + 9 + 1;
+}
+
+#[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, PartialEq, Eq)]
+pub enum MarketState {
+    Uninitialized,
+    Created,
+    Active,
+    Ended,
+}
+
+#[account]
+pub struct Market {
+    pub store: Pubkey,
+    pub selling_resource: Pubkey,
+    pub treasury_mint: Pubkey,
+    pub treasury_holder: Pubkey,
+    pub treasury_owner: Pubkey,
+    pub owner: Pubkey,
+    pub name: String,
+    pub description: String,
+    pub mutable: bool,
+    pub price: u64,
+    pub pieces_in_one_wallet: Option<u64>,
+    pub start_date: u64,
+    pub end_date: Option<u64>,
+    pub state: MarketState,
+}
+
+impl Market {
+    pub const LEN: usize = 8
+        + 32
+        + 32
+        + 32
+        + 32
+        + 32
+        + 32
+        + STRING_DEFAULT_SIZE * 4
+        + STRING_DEFAULT_SIZE * 4
+        + 1
+        + 8
+        + 9
+        + 8
+        + 9
+        + 1;
+}
+
+#[account]
+pub struct TradeHistory {
+    pub market: Pubkey,
+    pub wallet: Pubkey,
+    pub already_bought: u64,
+}
+
+impl TradeHistory {
+    pub const LEN: usize = 8 + 32 + 32 + 8;
+}
